backend/requirement.txt
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
what3words==3.1.1
